Fixed loading playerflag settings

+++ trunk/netpanzer/src/NetPanzer/Views/MainMenu/Multi/HostJoinTemplateView.cpp
@@ -90,7 +90,7 @@ void bReady()
     gameconfig->playername = PlayerNameView::playerName.getString();
 
     // Set the player flag.
-    gameconfig->playerflag = ( (short) playerFlagSelected );
+    gameconfig->playerflag = playerFlagSelected;
 
     if (gameconfig->hostorjoin == _game_session_join) {
         //winsock hack
